+++ src/constants/index.ts
@@ -1,5 +1,6 @@
 export { ErrorMessages } from "./error-messages.constants";
 export { propertyMetadata } from "./property-metadata.constants";
 export { ApiConfigService } from "./api-config.service";
+export { methodNameToEventName } from "./methodName-to-eventName.constants";
 
 export const DATE_FORMAT = "YYYY-MM-DD";